Fim do CRUD de Dados de Posição e população do banco de dados

+++ src/main/java/com/api/core/appl/dadosposicao/DadosPosicaoDTO.java
@@ -1,6 +1,7 @@
 package com.api.core.appl.dadosposicao;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 
 public class DadosPosicaoDTO implements Serializable {
@@ -13,9 +14,9 @@ public class DadosPosicaoDTO implements Serializable {
 
 	private Float velocidade;
 
-	private Double longitude;
+	private BigDecimal longitude;
 
-	private Double latitude;
+	private BigDecimal latitude;
 
 	private Boolean ignicao;
 
@@ -43,19 +44,19 @@ public class DadosPosicaoDTO implements Serializable {
 		this.velocidade = velocidade;
 	}
 
-	public Double getLongitude() {
+	public BigDecimal getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Double longitude) {
+	public void setLongitude(BigDecimal longitude) {
 		this.longitude = longitude;
 	}
 
-	public Double getLatitude() {
+	public BigDecimal getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Double latitude) {
+	public void setLatitude(BigDecimal latitude) {
 		this.latitude = latitude;
 	}
